Refactor notification listener
The notification listener does not need anymore to return capital
transfer information, since it's already included in the JobStart info
