+++ data-model/src/main/java/ch/prevo/open/node/data/provider/MatchNotificationListener.java
@@ -1,12 +1,11 @@
 package ch.prevo.open.node.data.provider;
 
-import ch.prevo.open.encrypted.model.CapitalTransferInformation;
-import ch.prevo.open.encrypted.model.TerminationMatchNotification;
 import ch.prevo.open.encrypted.model.CommencementMatchNotification;
+import ch.prevo.open.encrypted.model.TerminationMatchNotification;
 
 public interface MatchNotificationListener {
 
-    CapitalTransferInformation handleTerminationMatch(TerminationMatchNotification notification);
+    void handleTerminationMatch(TerminationMatchNotification notification);
 
     void handleCommencementMatch(CommencementMatchNotification notification);
